Comments added

+++ src/main/java/com/sumbaseassignment/Services/CustomerServices.java
@@ -5,6 +5,11 @@ import com.sumbaseassignment.Repository.CustomerRepository;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,9 +18,11 @@ import java.util.Optional;
 @Service
 public class CustomerServices {
 
+    // Autowired instance of CustomerRepository for accessing data layer.
     @Autowired
     private CustomerRepository customerRepository;
 
+    // Adds a new customer to the database.
     public Customer addCustomer(Customer customer) throws Exception{
         try {
             return customerRepository.save(customer);
@@ -25,6 +32,8 @@ public class CustomerServices {
         }
     }
 
+
+    // Retrieves a customer by their ID.
     public Customer getCustomerById(int customerId) throws  Exception{
 
         Optional<Customer> customer = customerRepository.findById(customerId);
@@ -36,14 +45,25 @@ public class CustomerServices {
     }
 
 
+    // Deletes a customer by their ID.
     public void deleteCustomerById(int customerId) throws Exception {
+
         try {
-            customerRepository.deleteById(customerId);
+            Optional<Customer> customer = customerRepository.findById(customerId);
+
+            if(customer.isEmpty()){
+                throw new Exception("Customer not found");
+            }
+            else {
+                customerRepository.deleteById(customerId);
+            }
+
         } catch (EmptyResultDataAccessException e) {
             throw new Exception("Invalid id: " + customerId, e);
         }
     }
 
+    // Updates details of an existing customer.
     public Customer updateCustomerDetails(int cutomerId, Customer updateCustomer) throws Exception {
 
         try {
@@ -70,8 +90,20 @@ public class CustomerServices {
         }
     }
 
-    public List<Customer> getAllCustomer() {
-        return customerRepository.findAll();
+
+
+    // Retrieves all customers with pagination.
+    public Page<Customer> getAllCustomers(Pageable pageable) {
+        return customerRepository.findAll(pageable);
     }
+
+    // Searches for customers based on a search term.
+    public List<Customer> searchCustomers(String searchTerm) {
+        return customerRepository.searchCustomers(searchTerm);
+    }
+
+
+
+
 
 }
